requirements.txt
openai==0.27.0
wechat_sdk==0.6.4
sanic==22.12.0
xmltodict~=0.9.2
PyYAML~=6.0
mysql-connector-python==8.0.32